Extract focusedRectInRoot function.
Test: No new tests required
